Add form

+++ src/Sylius/Bundle/AdminBundle/Controller/AvatarController.php
@@ -13,19 +13,18 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\AdminBundle\Controller;
 
-use Sylius\Bundle\CoreBundle\Doctrine\ORM\AvatarRepository;
 use Sylius\Component\Core\Model\AvatarImage;
+use Sylius\Component\Core\Repository\AvatarRepositoryInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\RouterInterface;
-use Webmozart\Assert\Assert;
 
 final class AvatarController
 {
-    /** @var AvatarRepository */
+    /** @var AvatarRepositoryInterface */
     private $avatarRepository;
 
     /** @var EngineInterface */
@@ -34,8 +33,11 @@ final class AvatarController
     /** @var RouterInterface */
     private $router;
 
-    public function __construct(RepositoryInterface $avatarRepository, EngineInterface $templatingEngine, RouterInterface $router)
-    {
+    public function __construct(
+        RepositoryInterface $avatarRepository,
+        EngineInterface $templatingEngine,
+        RouterInterface $router
+    ) {
         $this->avatarRepository = $avatarRepository;
         $this->templatingEngine = $templatingEngine;
         $this->router = $router;
@@ -45,8 +47,10 @@ final class AvatarController
     {
         /** @var AvatarImage $avatar */
         $avatar = $this->avatarRepository->findOneByOwner($id);
-        Assert::notNull($avatar);
-        $this->avatarRepository->remove($avatar);
+
+        if(null !== $avatar) {
+            $this->avatarRepository->remove($avatar);
+        }
 
         $url = $this->router->generate('sylius_admin_admin_user_update', ['id' => $id]);
